% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalRes.R
\name{calculer_residus_RLC}
\alias{calculer_residus_RLC}
\title{calcul integrale de la fonction de transfert}
\usage{
calculer_residus_RLC(R, L, C)
}
\arguments{
\item{R}{Resistance(en ohms)}

\item{L}{Inductance(en henrys)}

\item{C}{Capacité(en farads)}

\item{s}{a number}
}
\value{
number

a function
}
\description{
calcul integrale de la fonction de transfert
}
