% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonc_trans.R
\name{fonction_de_transfert_RLC}
\alias{fonction_de_transfert_RLC}
\title{fontion transfert d'un circuit RLC}
\usage{
fonction_de_transfert_RLC(R, L, C)
}
\arguments{
\item{R}{Résistance(en ohms)}

\item{L}{Inductance(en henrys)}

\item{C}{Capacité(en farads)}
}
\value{
fonction
}
\description{
fontion transfert d'un circuit RLC
}
