% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pol.R
\name{poles}
\alias{poles}
\title{Calcul des poles d'une fonction rationelle}
\usage{
poles(a, b, c)
}
\arguments{
\item{a}{numbers}

\item{b}{numbers}

\item{c}{numbers}
}
\value{
number
}
\description{
Calcul des poles d'une fonction rationelle
}
