% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pol_Zeros.R
\name{poles_zeros_RLC}
\alias{poles_zeros_RLC}
\title{calcul des poles et des zeros de la fonction de transfert}
\usage{
poles_zeros_RLC(R, L, C)
}
\arguments{
\item{R}{Résistance(en ohms)}

\item{L}{Inductance(en henrys)}

\item{C}{Capacité(en farads)}
}
\value{
number
}
\description{
calcul des poles et des zeros de la fonction de transfert
}
